Rename function for clarity
